+++ geofileops/_compat.py
@@ -20,6 +20,7 @@ GDAL_BASE_VERSION = version.parse(gdal.__version__.split("-")[0]).base_version
 GDAL_GTE_39 = version.parse(GDAL_BASE_VERSION) >= version.parse("3.9")
 GDAL_GTE_3101 = version.parse(GDAL_BASE_VERSION) >= version.parse("3.10.1")
 GDAL_GTE_311 = version.parse(GDAL_BASE_VERSION) >= version.parse("3.11")
+GDAL_GTE_3114 = version.parse(GDAL_BASE_VERSION) >= version.parse("3.11.4")
 
 GEOPANDAS_GTE_10 = version.parse(gpd.__version__) >= version.parse("1.0")
 GEOPANDAS_110 = version.parse(gpd.__version__) == version.parse("1.1.0")