+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add `concat` function (#746, #747)
 - Add `join` function (#751)
 - Add `add_columns` function (#768)
+- Add a general configuration option to remove slivers (#776)
 - Add `zip_geofile` and `unzip_geofile` functions (#754, #743)
 - Add `write_mode="append_add_fields"` option to `copy_layer` (#750)
 - Add an `include_duplicates` parameter to `union`, `intersection` and `identity` (#757)